requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
plotly>=5.20.0
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fear-greed>=0.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
